+++ src/gameLayer/randomStuff.hpp
@@ -3,6 +3,6 @@
 
 float getRandomFloat(std::ranlux24_base& rng, float min, float max);
 
-float getRandomInt(std::ranlux24_base& rng, int min, int max);
+int getRandomInt(std::ranlux24_base& rng, int min, int max);
 
 bool getRandomChance(std::ranlux24_base& rng, float chance);
